basic auth using clerc api+fonts
