+++ .env
@@ -1,0 +1,1 @@
+EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY=pk_test_c29saWQtY2F0ZmlzaC05OS5jbGVyay5hY2NvdW50cy5kZXYk
