--- app/_layout.tsx
@@ -1,13 +1,0 @@
-import { Stack } from "expo-router";
-
-export default function RootLayout() {
-  return (
-    <Stack screenOptions={{
-      headerShown:false
-    }}>
-      <Stack.Screen name="(tabs)" options={{
-        headerShown: false
-      }}/>
-    </Stack>
-  );
-}
